# src/main/java/com/scaffoldops/generatorapi/domain/model/DeploymentStatus.java
package com.scaffoldops.generatorapi.domain.model;

public enum DeploymentStatus {
    NOT_DEPLOYED,
    DEPLOYMENT_REQUESTED,
    DEPLOYING,
    DEPLOYED,
    UNDEPLOYMENT_REQUESTED,
    UNDEPLOYING,
    DEPLOYMENT_FAILED
}