# src/main/resources/openapi/generator-api.yaml
openapi: 3.0.3
info:
  title: ScaffoldOps Generator API
  description: API for submitting, tracking, deploying, and undeploying scaffold generation requests.
  version: "0.0.1-SNAPSHOT"
servers:
  - url: /api/generator/v1
    description: Base API path
tags:
  - name: GenerationRequest
    description: Public operations for scaffold generation requests.
  - name: InternalGenerationRequest
    description: Internal service-to-service operations. Not intended for use by end clients.
paths:
  /generation-requests:
    post:
      tags:
        - GenerationRequest
      operationId: createGenerationRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - GenerationRequest
      operationId: listGenerationRequests
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: template
          in: query
          required: false
          schema:
            type: string
        - name: generationStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GenerationStatus'
        - name: deploymentStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeploymentStatus'
        - name: deploymentTarget
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeploymentTarget'
        - name: database
          in: query
          required: false
          schema:
            type: boolean
        - name: restApi
          in: query
          required: false
          schema:
            type: boolean
        - name: security
          in: query
          required: false
          schema:
            type: boolean
        - name: messaging
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenerationRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /generation-requests/{id}:
    get:
      tags:
        - GenerationRequest
      operationId: getGenerationRequestById
      parameters:
        - $ref: '#/components/parameters/GenerationRequestId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - GenerationRequest
      operationId: deleteGenerationRequestById
      parameters:
        - $ref: '#/components/parameters/GenerationRequestId'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /generation-requests/{id}/deploy:
    post:
      tags:
        - GenerationRequest
      operationId: deployGenerationRequest
      parameters:
        - $ref: '#/components/parameters/GenerationRequestId'
      responses:
        '202':
          description: Deployment requested
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /generation-requests/{id}/deployment:
    delete:
      tags:
        - GenerationRequest
      operationId: undeployGenerationRequest
      parameters:
        - $ref: '#/components/parameters/GenerationRequestId'
      responses:
        '202':
          description: Undeployment requested
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /internal/generation-requests/{requestId}/generation-status:
    patch:
      tags:
        - InternalGenerationRequest
      summary: Update generation lifecycle status
      description: Internal callback used by generator-worker to report generation progress and results.
      operationId: updateGenerationRequestGenerationStatus
      parameters:
        - $ref: '#/components/parameters/InternalGenerationRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationStatusUpdateRequest'
      responses:
        '204':
          description: Generation status updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /internal/generation-requests/{requestId}/deployment-status:
    patch:
      tags:
        - InternalGenerationRequest
      summary: Update deployment lifecycle status
      description: Internal callback used by deployment-worker to report deployment progress and results.
      operationId: updateGenerationRequestDeploymentStatus
      parameters:
        - $ref: '#/components/parameters/InternalGenerationRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentStatusUpdateRequest'
      responses:
        '204':
          description: Deployment status updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  parameters:
    GenerationRequestId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InternalGenerationRequestId:
      name: requestId
      in: path
      required: true
      description: Generation request identifier.
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Invalid lifecycle state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    DeploymentTarget:
      type: string
      enum:
        - KUBERNETES
    GenerationStatus:
      type: string
      enum:
        - RECEIVED
        - GENERATING
        - GENERATED
        - GENERATION_FAILED
    DeploymentStatus:
      type: string
      enum:
        - NOT_DEPLOYED
        - DEPLOYMENT_REQUESTED
        - DEPLOYING
        - DEPLOYED
        - UNDEPLOYMENT_REQUESTED
        - UNDEPLOYING
        - DEPLOYMENT_FAILED
    CreateGenerationRequestRequest:
      type: object
      required:
        - name
        - template
        - deploymentTarget
      properties:
        name:
          type: string
          minLength: 1
        template:
          type: string
          minLength: 1
        database:
          type: boolean
        restApi:
          type: boolean
        security:
          type: boolean
        messaging:
          type: boolean
        deploymentTarget:
          $ref: '#/components/schemas/DeploymentTarget'
    GenerationStatusUpdateRequest:
      type: object
      required:
        - generationStatus
      properties:
        generationStatus:
          type: string
          description: Current generation status reported by generator-worker.
          enum:
            - GENERATING
            - GENERATED
            - GENERATION_FAILED
        message:
          type: string
          description: Optional progress or failure detail.
        artifactRef:
          type: string
          description: Optional reference to the generated artifact.
    DeploymentStatusUpdateRequest:
      type: object
      required:
        - deploymentStatus
      properties:
        deploymentStatus:
          type: string
          description: Current deployment status reported by deployment-worker.
          enum:
            - DEPLOYING
            - DEPLOYED
            - UNDEPLOYING
            - NOT_DEPLOYED
            - DEPLOYMENT_FAILED
        message:
          type: string
          description: Optional progress or failure detail.
        imageRef:
          type: string
          description: Optional reference to the deployed container image.
    GenerationRequestResponse:
      type: object
      required:
        - id
        - name
        - template
        - deploymentTarget
        - generationStatus
        - deploymentStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        template:
          type: string
        database:
          type: boolean
        restApi:
          type: boolean
        security:
          type: boolean
        messaging:
          type: boolean
        deploymentTarget:
          $ref: '#/components/schemas/DeploymentTarget'
        generationStatus:
          $ref: '#/components/schemas/GenerationStatus'
        deploymentStatus:
          $ref: '#/components/schemas/DeploymentStatus'
        message:
          type: string
          nullable: true
        artifactRef:
          type: string
          nullable: true
        imageRef:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - message
        - timestamp
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time